+++ src/lib/components/Form/DumbInput.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
 	import { cn } from '$lib/utils';
 	import { Label, type LabelRootProps } from 'bits-ui';
-	import { type Icon as IconType } from 'lucide-svelte';
+	import { type IconProps, type Icon as IconType } from 'lucide-svelte';
+	import type { Snippet } from 'svelte';
 	import type { HTMLInputAttributes } from 'svelte/elements';
 	import { tv } from 'tailwind-variants';
 
@@ -20,11 +21,13 @@
 	});
 
 	interface Props extends HTMLInputAttributes {
-		label: string;
+		label?: string | Snippet;
 		labelProps?: Omit<LabelRootProps, 'for'>;
 		icon?: typeof IconType;
+		iconProps?: IconProps;
 		errors?: string[];
 		variant?: 'default' | 'ghost';
+		inputRef?: HTMLInputElement;
 	}
 
 	let {
@@ -36,6 +39,8 @@
 		icon: Icon,
 		errors = [],
 		variant = 'default',
+		inputRef = $bindable(),
+		iconProps,
 		...props
 	}: Props = $props();
 	let className = $derived(
@@ -50,16 +55,28 @@
 	);
 </script>
 
-<Label.Root {...labelProps} for={name} class="relative flex w-full flex-col gap-xs">
-	<span class="font-bold">{label}</span>
+<div class="relative flex w-full select-text flex-col gap-xs">
+	<Label.Root {...labelProps} for={name} class={cn(label == null && 'hidden', labelProps?.class)}>
+		{#if typeof label === 'string'}
+			<span class="font-bold">{label}</span>
+		{:else if label != null}
+			{@render label()}
+		{/if}
+	</Label.Root>
 	<span class="relative">
-		<input {...props} {type} {name} class={className} bind:value />
+		<input {...props} {type} {name} class={className} bind:value bind:this={inputRef} />
 		{#if Icon != null}
-			<Icon class="absolute right-md top-1/2 size-5 -translate-y-1/2 text-muted-foreground" />
+			<Icon
+				{...iconProps}
+				class={cn(
+					'absolute right-md top-1/2 size-5 -translate-y-1/2 text-muted-foreground transition-all duration-300',
+					iconProps?.class
+				)}
+			/>
 		{/if}
 	</span>
 
 	{#each errors as error}
 		<p class="error">{error}</p>
 	{/each}
-</Label.Root>
+</div>